+++ src/main/java/mcheli/aircraft/MCH_RenderBaseVehicle.java
@@ -497,7 +497,7 @@ public abstract class MCH_RenderBaseVehicle extends W_Render {
       return prevRot + (rot - prevRot) * tickTime;
    }
 
-   public void renderDebugHitBox(MCH_EntityBaseVehicle e, double x, double y, double z, float yaw, float pitch) {
+   public void renderDebugHitBox(MCH_EntityBaseVehicle e, double x, double y, double z, float yaw, float pitch, float roll) {
       MCH_Config var10000 = MCH_MOD.config;
       if(MCH_Config.TestMode.prmBool && debugModel != null) {
          GL11.glPushMatrix();
@@ -508,6 +508,7 @@ public abstract class MCH_RenderBaseVehicle extends W_Render {
          GL11.glPopMatrix();
          GL11.glPushMatrix();
          GL11.glTranslated(x, y, z);
+         boolean drawObb = Minecraft.getMinecraft().isSingleplayer() && MCH_Config.DebugDrawOBB.prmBool;
          MCH_BoundingBox[] arr$ = e.getCalculatedExtraBoundingBoxes();
          int len$ = arr$.length;
 
@@ -515,6 +516,11 @@ public abstract class MCH_RenderBaseVehicle extends W_Render {
             MCH_BoundingBox bb = arr$[i$];
             GL11.glPushMatrix();
             GL11.glTranslated(bb.rotatedOffset.xCoord, bb.rotatedOffset.yCoord, bb.rotatedOffset.zCoord);
+            if(drawObb) {
+               GL11.glRotatef(yaw, 0.0F, -1.0F, 0.0F);
+               GL11.glRotatef(pitch, 1.0F, 0.0F, 0.0F);
+               GL11.glRotatef(roll, 0.0F, 0.0F, 1.0F);
+            }
             GL11.glPushMatrix();
             GL11.glScalef(bb.width, bb.height, bb.width);
             this.bindTexture("textures/bounding_box.png");